backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
chromadb==0.4.18
langchain==0.1.0
langchain-openai==0.0.2
openai>=1.99.5
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken>=0.7.0
pdfplumber==0.10.3
anyio>=4.8.0
sentence-transformers==2.2.2
